requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9

# Machine Learning & Boosting
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2

# Deep Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.1

# Data Processing & Math
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Visualization & Image Processing
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
opencv-python==4.9.0.80

# Utilities
kaggle==1.6.3
python-dotenv==1.0.1